requirements.txt
Flask==3.1.0
requests==2.32.3
oauth2client==4.1.3
#uwsgi==2.0.20  # for appengine, hide for container
websockets==15.0.1
asyncio==3.4.3
# grpc
grpcio==1.71.0
grpcio-tools==1.71.0
protobuf==5.29.4
Werkzeug==3.1.3
functions-framework==3.*
google-api-python-client==2.166.0
google-cloud-dialogflow-cx==2.1.0
#azure bot
aiohttp==3.13.2
botbuilder-core==4.17.0
botbuilder-schema==4.17.0